+++ src/seedu/addressbook/data/person/Contact.java
@@ -1,0 +1,26 @@
+package seedu.addressbook.data.person;
+
+public class Contact {
+	
+	public String value;
+    protected boolean isPrivate;
+    
+    public Contact(String value, boolean isPrivate) {
+    	this.value = value;
+    	this.isPrivate = isPrivate;
+    }
+    
+    @Override
+    public String toString() {
+        return value;
+    }
+    
+    @Override
+    public int hashCode() {
+        return value.hashCode();
+    }
+
+    public boolean isPrivate() {
+        return isPrivate;
+    }
+}
